fix videoId type in film and episode entities

+++ src/episode/entities/episode.entity.ts
@@ -89,7 +89,7 @@ export class EpisodeEntity {
   series: SeriesEntity;
 
   @FilterableField({ nullable: true })
-  @Column({ nullable: true, type: 'int8' })
+  @Column({ nullable: true, type: 'int4' })
   @Index({ where: 'video_id IS NOT NULL' })
   videoId?: number;
